requirements.txt
langchain
langchain-community
langgraph
ollama
langchain-ollama
chatterbox-tts 
fastapi
uvicorn
pygame
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client